// src/app/dashboard/interfaces/votante-response.interface.ts
export interface VotanteResponse {
    id:            number;
    nombrepadre:   string;
    idpadre:       number;
    nombre:        string;
    telefono:      string;
    inefrente:     string;
    ocr:           string;
    fecha:         string;
    domilio:       string;
    referenciadom: string;
    colonia:       string;
    seccion:       string;
    programas:     string;
    observaciones: string;
    traslado:      boolean;
}